requirements.txt
langchain== 0.0.154
PyPDF2 ==3.0.1
python-dotenv == 1.1.0
stremlit == 1.18.1
faiss-cpu == 1.7.4
